+++ llm_guard/input_scanners/__init__.py
@@ -4,6 +4,7 @@ from .ban_substrings import BanSubstrings
 from .ban_topics import BanTopics
 from .code import Code
 from .prompt_injection import PromptInjection
+from .prompt_injection_v2 import PromptInjectionV2
 from .secrets import Secrets
 from .sentiment import Sentiment
 from .token_limit import TokenLimit
@@ -15,6 +16,7 @@ __all__ = [
     "BanTopics",
     "Code",
     "PromptInjection",
+    "PromptInjectionV2",
     "Secrets",
     "Sentiment",
     "TokenLimit",